Revert "GG-37103 TC check."
This reverts commit de5eb3d46009f41d5b76d6b6fd5d29ef737b64dd.

+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/database/H2TreeIndexBase.java
@@ -116,12 +116,11 @@ public abstract class H2TreeIndexBase extends GridH2IndexBase {
         if (propSize == 0)
             return 0;
 
-        // TODO rework
-        long size = 0;
+        int size = 0;
 
         for (InlineIndexColumn idxHelper : inlineIdxs) {
             // for variable types - default variable size, for other types - type's size + type marker
-            long sizeInc = idxHelper.size() < 0 ? IGNITE_VARIABLE_TYPE_DEFAULT_INDEX_SIZE : idxHelper.size() + 1;
+            int sizeInc = idxHelper.size() < 0 ? IGNITE_VARIABLE_TYPE_DEFAULT_INDEX_SIZE : idxHelper.size() + 1;
 
             fixedSize &= idxHelper.size() != -1;
 
@@ -133,7 +132,7 @@ public abstract class H2TreeIndexBase extends GridH2IndexBase {
 
                     if (m.find())
                         // if column has defined length we use it as default + 3 bytes for the inner info of the variable type
-                        sizeInc = Long.parseLong(m.group(1)) + 3;
+                        sizeInc = Integer.parseInt(m.group(1)) + 3;
                 }
             }
 
@@ -152,13 +151,13 @@ public abstract class H2TreeIndexBase extends GridH2IndexBase {
                         "This will lead to wasting of space inside index pages. Ignoring " +
                         "[index=" + name + ", explicitInlineSize=" + cfgInlineSize + ", realInlineSize=" + size + ']');
 
-                return (int) size;
+                return size;
             }
 
             return cfgInlineSize;
         }
 
-        return (int)Math.min(PageIO.MAX_PAYLOAD_SIZE, size);
+        return Math.min(PageIO.MAX_PAYLOAD_SIZE, size);
     }
 
     /**
